Refactor de los manejadores a configuradores
